requirements.txt
yt-dlp
google.generativeai
streamlit